Fix a version-byte trap, and turn two wire-format claims into assertions
The wire-format round walked every length prefix, every optional field and the
whole signature coverage and could not construct a second spelling of one
envelope or a field the signature misses. What it found was three claims and a
type.

MAX_WIRE_CHARS justified itself with "the send side caps bundles at 4096". That
cap is real and is reached only from the invite-only path:
checkMessageLengthForEncodingMethod returns inside its isPreKeyResponse branch
before the message branches, so 4096 and the message limits are mutually
exclusive and neither applies to a wire text carrying a bundle AND a ciphertext.
That is the ordinary shape whenever the signed pre-key rotation falls due, and
this repo's own figure for it is 5584 characters. Believing the sentence would
mean tightening the constant toward 4096 and refusing every rotation-attached
message on the receiver, silently, for an ordinary send. Now an assertion: that
envelope is larger than the invite cap and smaller than the decoder's.

VERSION was a byte compared against a u8. Any future value from 128 up
sign-extends, so the comparison fails for every envelope - including ones the
encoder has just written, since write(int) emits the low eight bits either way.
The decoder would refuse its own output on a routine bump. Nothing is wrong at 2
and nothing would be up to 127; the trap springs on the bump after next. It is
an int now, and not fixed by masking on the write side, which would make the
wire byte right and leave the comparison signed.

And an invariant the encoder depends on that nothing wrote down: an account
carries its device id twice, the two envelope builders read different copies,
and encode refuses anything outside libsignal's range. They agree only because
both new Account(...) sites pass an already-folded value - while the store still
writes a DEVICE_ID row nothing reads, holding the unfolded nextInt(10000) value
on a 0.1.5 store, looking exactly like the field a third caller would reach for.

That one is asserted after a REAL load. I wrote it first against a fresh
in-memory account, where it passed with the invariant deliberately broken,
because that path never reaches the constructor that makes it true.

+++ app/src/main/java/com/amnesica/kryptey/inputmethod/signalprotocol/encoding/BinaryEnvelope.java
@@ -85,7 +85,21 @@ public final class BinaryEnvelope {
    * Mixed-version compatibility is not a requirement here, so the decoder refuses version one
    * outright rather than accepting an unsigned bundle from an older build.
    */
-  static final byte VERSION = 2;
+  /**
+   * An {@code int}, not a {@code byte}, and the difference is a trap rather than a preference.
+   *
+   * <p>{@code decode} compares this against {@code c.u8("version")}, which returns 0 to 255. As a
+   * {@code byte} any future value from 128 up sign-extends to a negative number, so the comparison
+   * fails for every envelope - including ones this encoder has just written, since
+   * {@code write(int)} emits the low eight bits either way. The decoder would refuse its own output
+   * and every message from every peer would read as "not a valid envelope", on a routine version
+   * bump. Nothing is wrong at 2, and nothing would be up to 127; the trap only springs on the
+   * bump after next, which is exactly the kind that gets made quickly.
+   *
+   * <p>Not fixed by masking on the write side. That makes the wire byte right and leaves the
+   * comparison signed, so the refusal stays and the cause moves.
+   */
+  static final int VERSION = 2;
   private static final int FLAG_PRE_KEY_RESPONSE = 0x01;
   private static final int FLAG_CIPHERTEXT = 0x02;
